--- a/core/register_core_types.cpp
+++ b/core/register_core_types.cpp
@@ -100,8 +100,6 @@ extern void register_variant_methods();
 extern void unregister_variant_methods();
 
 void register_core_types() {
-	MemoryPool::setup();
-
 	StringName::setup();
 
 	register_global_constants();
@@ -316,6 +314,5 @@ void unregister_core_types() {
 	ResourceCache::clear();
 	CoreStringNames::free();
 	StringName::cleanup();
-
-	MemoryPool::cleanup();
+	MemoryPool::report_leaks();
 }