… core/pool_allocator.cpp
… core/pool_allocator.h
… editor/plugins/spatial_editor_plugin.cpp
… modules/gdnative/gdnative/array.cpp
… scene/2d/cpu_particles_2d.cpp
… scene/3d/cpu_particles.cpp
… scene/3d/mesh_instance.cpp
… scene/3d/room_manager.cpp
… scene/animation/animation_tree_player.cpp
… scene/main/resource_preloader.cpp
… scene/resources/mesh.cpp
… scene/resources/multimesh.cpp
… scene/resources/occluder_shape_polygon.cpp
… servers/physics/shape_sw.cpp
… servers/physics_2d/shape_2d_sw.cpp