Modify PoolVector internals to use LocalVector internally, and remove the hard limit on the number of PoolVectors.

While in master, PoolVector has been replaced wholesale with Vector, in order to minimize disruption to existing codebase, I have been testing making PoolVector a simple wrapper.

Fixes #37154
Fixes #53665
Alternative to #54263

## Notes
* Still a WIP, particularly I need to check the locking.
* This version unlike the original `PoolVector` allocates no memory up front, and has no hard limits to the number of `PoolVector`s. This means memory RAM use is typically lower than before, and is probably a better approach to the limit than increasing it and increasing the fixed RAM cost. 
* I first tested this wrapping `Vector`, however this duplicate the COW functionality in `PoolVector`, so I'm trying `LocalVector` which has no COW functionality, and it seems to be working fine so far.
* Any locking / multithread safety is handled by `PoolVector` and the `Access`, as `LocalVector` is not thread safe.
* If we are happy with this we may decide to put the `PoolVector_Alloc`s into one of the CommonPools, using e.g. `PagedAllocator`, rather than make the dynamic allocations using malloc (although of course the vector allocations themselves for resizing will still use malloc).

<!--
Pull requests should always be made for the `master` branch first, as that's
where development happens and the source of all future stable release branches.

Relevant fixes are cherry-picked for stable branches as needed.

Do not create a pull request for stable branches unless the change is already
available in the `master` branch and it cannot be easily cherry-picked.
Alternatively, if the change is only relevant for that branch (e.g. rendering
fixes for the 3.2 branch).
-->