@jordo 2022-03-27
@lawnjelly @reduz can we setup a discussion on contributors chat for this stuff?   I feel like I'm really lost with whats going on with this stuff as of late, and I'm having a hard time understanding what changes are being made, what the problems are, what are the design goals, etc.  I'm not opposed to any of this, and welcome all and any efforts to improve (there definitely needs to be improvement), but the changes are coming in hot and fast with not much formal discussion or technical proposals of what exactly the issues are, how any PR and/or refactoring directly addresses them, and benchmarks to examine .

The core fundamentals & data structures of the engine are in dire need of additional analysis & scrutiny, unit testing & benchmarks, and technical documentation that can be understood a group of contributors.  This is a key area that I have, over the past year, spent too many hours modifying and working around to meet performance requirements of a game where we needed to be performance first.  

1) - Can we define technically, what the design goals of `PoolAllocator` were, and how it accomplished them?  Because this PR removes it completely.  From what I can tell on a first glance, it's a pre-allocated block, which is/was used as a generic memory `pool` with a limited amount of compaction and resizing capabilities per `Entry`.

2) - Since this PR maintains the data structure called `PoolVector`, in 3.X we have `Vector`, `LocalVector`, and `PoolVector` (not that we don't already have them already).  But what exactly is PoolVector, and what does it do?  Is it documented anywhere?  Is it effectively now just a reference counted, COW, `LocalVector`?  Is it even supposed to be thread safe, or not?  Does it have stable element references on iteration?  etc.  

The second comment above is more a general thought about the core structures of the engine, but it illustrates a point that none of this is really well communicated or formally documented anywhere.  

As a contrast, we've moved almost all of our performance critical  Dictionary & Hashmap operations to https://github.com/martinus/robin-hood-hashing (either it's flat_map, or node based memory backing layouts), and seen absolutely ridiculous performance improvements across the board and almost everywhere.  This single data structure's test dir has an incredible number of unit tests and benchmarks (https://github.com/martinus/robin-hood-hashing/tree/master/src/test/unit).  I just think it's a great example, that the foundational building blocks of a project as large as godot, especially data structures, need to have proper documentation, tests, and benchmark suites.  Otherwise its very difficult to understand what's changing, what the goals are, and try to objectively determine what results in the best outcome performance wise.  

With regards to the direct questions and comments that @lawnjelly brought up in this PR, there's a lot.  Would we be able to schedule a contributors discussion and perhaps even post a more formal write up with what the problems are (other than just linked issue #'s), and what we are changing and why?  I think that would be really helpful.