@lawnjelly 2022-04-04
Some performance timings of different release compiles of `PoolVector`, in milliseconds, smaller times is therefore faster.

These are in gdscript so not super gold standard, because there will be a large fixed cost due to the scripting, but easier to rig up, and give us a ballpark idea of whether various approaches are faster or slower.

Here's the project:
[PoolVectorBenchmark.zip](https://github.com/godotengine/godot/files/8408764/PoolVectorBenchmark.zip)

### 1024x1024 push backs, 10 runs
Old PoolVector : 1899, 1826, 1872, 1808
This PR, global lock : 1431, 1393, 1402, 1378
This PR, lock per alloc : 1579, 1583, 1639, 1593

### 1024x1024 sets, 10 runs
Old PoolVector : 1116, 1099, 1072, 1078
This PR, global lock : 1107, 1057, 1053, 1065
This PR, lock per alloc : 1173, 1185, 1239, 1198